separate out cwl encoder into its own package
