--- cwl/src/main/scala/cwl/CwlEncoder.scala
@@ -1,11 +1,0 @@
-package cwl
-
-import shapeless._
-
-object CwlEncoder extends Poly1 {
-  implicit def workflow = at[Workflow]{ CwlCodecs.encodeCwlWorkflow }
-
-  implicit def commandLineTool = at[CommandLineTool]{ CwlCodecs.encodeCwlCommandLineTool }
-
-  implicit def expressionTool = at[ExpressionTool]{ CwlCodecs.encodeCwlExpressionTool }
-}
